requirements.txt
discord.py>=2.0.0
APScheduler>=3.10.1
python-dotenv>=1.0.0
aiosqlite>=0.19.0 